refactor: 파일들로 클래스들과 해당 테스트들을 분리한다.
