+++ src/main/kotlin/minesweeper/domain/Board.kt
@@ -1,0 +1,5 @@
+package minesweeper.domain
+
+sealed interface Board {
+    val cells: Map<Coordinate, Cell>
+}
