+++ src/main/kotlin/minesweeper/domain/Cell.kt
@@ -1,0 +1,3 @@
+package minesweeper.domain
+
+sealed interface Cell
