+++ src/main/kotlin/minesweeper/domain/EmptyCell.kt
@@ -1,0 +1,7 @@
+package minesweeper.domain
+
+sealed interface EmptyCell : Cell
+
+data object ClosedEmptyCell : EmptyCell
+
+data object OpenedEmptyCell : EmptyCell
