+++ src/main/kotlin/minesweeper/domain/Game.kt
@@ -1,0 +1,5 @@
+package minesweeper.domain
+
+sealed interface Game {
+    val board: Board
+}
