+++ src/main/kotlin/minesweeper/domain/PlayableBoard.kt
@@ -1,9 +1,5 @@
 package minesweeper.domain
 
-sealed interface Board {
-    val cells: Map<Coordinate, Cell>
-}
-
 class PlayableBoard(
     override val cells: Map<Coordinate, Cell>,
 ) : Board {
@@ -100,28 +96,6 @@ class PlayableBoard(
         newCells
             .values
             .count { it is ClosedEmptyCell } == 0
-}
-
-sealed interface CompletedBoard : Board
-
-class PlayerWonBoard(
-    override val cells: Map<Coordinate, Cell>,
-) : CompletedBoard {
-    init {
-        require(!cells.values.any { it is ClosedEmptyCell }) {
-            "승리한 게임은 지뢰를 제외한 모든 칸을 열어야 합니다."
-        }
-    }
-}
-
-class MineDetonatedBoard(
-    override val cells: Map<Coordinate, Cell>,
-) : CompletedBoard {
-    init {
-        require(cells.values.any { it is DetonatedMineCell }) {
-            "폭발한 지뢰가 있어야 합니다."
-        }
-    }
 }
 
 internal fun <K, V> Map<K, V>.replacing(
